feat: make gestures available on control area
Problem: The script decides which midi messages should be
forwarded to the instrument inputs using the translation table.
Sadly, this way the script can only capture note on and note off
messages.

Solution: Switch back to use channel to forward messages to the
instruments, as it was before, but retain the translation table to keep
the no-overlap row offset.

After this the collapse control surface does not work, the split needs
to be collapsed sending midi controls to the Linnstrument.